Ajustes a editar y eliminar

+++ listar.php
@@ -21,8 +21,18 @@ $result = pg_query($con, $sql);
 </head>
 
 <body>
+    <div class="d-flex flex-column flex-md-row align-items-center p-3 px-md-4 mb-3 bg-white border-bottom box-shadow">
+      <h5 class="my-0 mr-md-auto font-weight-normal"><img src="index2.png" style="width: 30px; position: absolute;"> <span style="position: relative; left: 35px;">Index</span></h5>
+      <nav class="my-2 my-md-0 mr-md-3">
+        <a class="p-2 text-dark" href="./index.php">Registrar</a>
+      </nav>
+    </div>
+
+    <div class="container px-3 py-3 pt-md-5 pb-md-4 mx-auto text-center">
+      <h1 class="display-4">Listado de Registros</h1>
+    </div>
+
     <div class="container">
-        <h2 class="mt-5 mb-4">Lista de Registros</h2>
         <table class="table">
             <thead>
                 <tr>
@@ -31,6 +41,8 @@ $result = pg_query($con, $sql);
                     <th>Apellidos</th>
                     <th>Dirección</th>
                     <th>Celular</th>
+                    <th></th>
+                    <th></th>
                 </tr>
             </thead>
             <tbody>
@@ -41,13 +53,20 @@ $result = pg_query($con, $sql);
                         <td><?php echo $row['apellido']; ?></td>
                         <td><?php echo $row['direccion']; ?></td>
                         <td><?php echo $row['celular']; ?></td>
-                        <td><a href="eliminar.php?id=<?php echo $row['idpersona']; ?>" class="btn btn-danger btn-sm">Eliminar</a></td>
+                        <td><a href="eliminar.php?id=<?php echo $row['idpersona']; ?>" class="btn btn-danger btn-sm" onclick="return confirmarEliminacion()">Eliminar</a></td>
                         <td><a href="editar.php?id=<?php echo $row['idpersona']; ?>" class="btn btn-primary btn-sm">Editar</a></td>
                     </tr>
                 <?php } ?>
             </tbody>
         </table>
     </div>
+
+    <script>
+        function confirmarEliminacion() {
+            return confirm('¿Estás seguro de que quieres eliminar este registro?');
+        }
+    </script>
+
     <!-- Bootstrap core JavaScript -->
     <script src="https://code.jquery.com/jquery-3.2.1.slim.min.js" integrity="sha384-KJ3o2DKtIkvYIK3UENzmM7KCkRr/rE9/Qpg6aAZGJwFDMVNA/GpGFF93hXpG5KkN" crossorigin="anonymous"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.12.9/umd/popper.min.js" integrity="sha384-ApNbgh9B+Y1QKtv3Rn7W3mgPxhU9K/ScQsAP7hUibX39j7fakFPskvXusvfa0b4Q" crossorigin="anonymous"></script>
